+++ SourceryRuntime/Sources/Composer.swift
@@ -691,6 +691,7 @@ public enum Composer {
             }
 
             baseType.based.keys.forEach { type.based[$0] = $0 }
+            baseType.basedTypes.forEach { type.basedTypes[$0.key] = $0.value }
             baseType.inherits.forEach { type.inherits[$0.key] = $0.value }
             baseType.implements.forEach { type.implements[$0.key] = $0.value }
 
@@ -709,6 +710,8 @@ public enum Composer {
                 // TODO: associated types?
                 type.implements[globalName] = baseType
             }
+
+            type.basedTypes[globalName] = baseType
         }
     }
 }